Game driver + packages refactor
