--- src/main/java/ar/fiuba/tdd/tp/GameBuilder.java
@@ -1,7 +1,0 @@
-package ar.fiuba.tdd.tp;
-
-import ar.fiuba.tdd.tp.model.Game;
-
-public interface GameBuilder {
-    Game build();
-}
