+++ src/main/java/ar/fiuba/tdd/tp/driver/GameDriver.java
@@ -1,0 +1,7 @@
+package ar.fiuba.tdd.tp.driver;
+
+public interface GameDriver {
+    void initGame(String jarPath);
+
+    String sendCommand(String cmd);
+}
