+++ src/main/java/ar/fiuba/tdd/tp/model/Game.java
@@ -2,4 +2,8 @@ package ar.fiuba.tdd.tp.model;
 
 // TODO: Dummy implementation just to make the project compile. Each group should use its own game class.
 public class Game {
+
+    public String getName() {
+        return "Some game";
+    }
 }
