+++ src/main/java/ar/fiuba/tdd/tp/model/GameBuilder.java
@@ -1,0 +1,5 @@
+package ar.fiuba.tdd.tp.model;
+
+public interface GameBuilder {
+    Game build();
+}
